# docs/swagger.json
{
  "openapi": "3.0.0",
  "components": {
    "schemas": {
      "ApiErrors": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "ProdutoRequest": {
        "type": "object",
        "required": ["descricao", "situacao"],
        "properties": {
          "descricao": {
            "type": "string",
            "pattern": "\\S"
          },
          "situacao": {
            "$ref": "#/components/schemas/SituacaoProdutoEnum"
          }
        }
      },
      "ProdutoResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "descricao": {
            "type": "string"
          },
          "situacao": {
            "$ref": "#/components/schemas/SituacaoProdutoEnum"
          }
        }
      },
...
      "SituacaoProdutoEnum": {
        "type": "string",
        "enum": ["ATIVO", "INATIVO"]
      }
    }
  },
  "paths": {
...
    "/api/produto": {
      "get": {
        "parameters": [
          {
            "name": "descricao",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "situacao",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de Produtos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProdutoResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrors"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrors"
                }
              }
            }
          }
        },
        "summary": "Get All",
        "tags": ["Produto Controller"]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProdutoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Produto criado com sucesso"
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrors"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrors"
                }
              }
            }
          }
        },
        "summary": "Create",
        "tags": ["Produto Controller"]
      }
    },
    "/api/produto/{id}": {
      "put": {
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProdutoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Produto atualizado com sucesso"
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrors"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrors"
                }
              }
            }
          }
        },
        "summary": "Update",
        "tags": ["Produto Controller"]
      },
      "get": {
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produto encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProdutoResponse"
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrors"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrors"
                }
              }
            }
          }
        },
        "summary": "Get By Id",
        "tags": ["Produto Controller"]
      },
      "delete": {
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produto excluído com sucesso"
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrors"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrors"
                }
              }
            }
          }
        },
        "summary": "Delete",
        "tags": ["Produto Controller"]
      }
    },
    "/api/produto/{id}/situacao": {
      "patch": {
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Situação do produto atualizada com sucesso"
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrors"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrors"
                }
              }
            }
          }
        },
        "summary": "Update Situacao",
        "tags": ["Produto Controller"]
      }
    }
  },
  "info": {
    "title": "Fisctrack API",
    "version": "1.0",
    "description": "API para gerenciamento de notas fiscais",
    "contact": {
      "name": "Suporte",
      "email": "wiriswernek@gmail.com",
      "url": "https://wiriswernek.tech"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ]
}